Revert "CSS update"
