+++ script.js
@@ -1,8 +1,4 @@
-import {drawGraphic, replaceMathFunctions, isEvaluable, isNotEmpty} from './functions.js';
-
 const elementList = document.getElementById('elementList');
-const graphicDialog = document.getElementById('graphicDialog');
-
 
 fetch('http://localhost:8080/api/functions')
   .then(response => response.json())
@@ -13,11 +9,6 @@ fetch('http://localhost:8080/api/functions')
       const updateButton = document.createElement('button');
       const drawButton = document.createElement('button');
       const deleteButton = document.createElement('button');
-      const errorElement = document.createElement('p')
-      const errorText = document.createTextNode("");
-      
-      errorElement.appendChild(errorText);
-      errorElement.classList.add('errorTextStyle');
       
       input.type = 'text';
       input.value = element.function_text;
@@ -37,19 +28,14 @@ fetch('http://localhost:8080/api/functions')
           .catch(error => console.error(error));
           location.reload();
         } else {
-          errorText.textContent = "*Введена функція невірна!";
+          output.innerHTML = "Введена функція невірна!";
         }
       });
       
       drawButton.textContent = 'Draw Graph';
       drawButton.addEventListener('click', () => {
         const value = element.function_text;
-        if (isEvaluable(replaceMathFunctions(value))) {
-          drawGraphic(replaceMathFunctions(value));
-          graphicDialog.style.display = "block";
-        } else {
-          errorText.textContent = "*Введена функція невірна!";
-        }
+        drawGraphic(replaceMathFunctions(value));
       });
 
       deleteButton.textContent = 'Delete';
@@ -75,13 +61,73 @@ fetch('http://localhost:8080/api/functions')
       li.appendChild(updateButton);
       li.appendChild(drawButton);
       li.appendChild(deleteButton);
-      li.appendChild(errorElement);
       elementList.appendChild(li);
     });
   })
   .catch(error => console.error(error));
 
 
+function drawGraphic(value) {
+  console.log(value);
+  const canvas = document.getElementById("myCanvas");
+  const ctx = canvas.getContext("2d");
 
+  const xMin = -10; // minimum x value
+  const xMax = 10; // maximum x value
+  const step = 0.1; // step size for x
+  const scale = 30; // pixels per unit
+  const originX = canvas.width / 2; // x coordinate of the origin
+  const originY = canvas.height / 2; // y coordinate of the origin
+
+  // draw x and y axes
+  ctx.beginPath();
+  ctx.moveTo(0, originY);
+  ctx.lineTo(canvas.width, originY);
+  ctx.moveTo(originX, 0);
+  ctx.lineTo(originX, canvas.height);
+  ctx.stroke();
+
+  // draw x and y axis labels
+  ctx.font = "12px Arial";
+  ctx.fillText("x", canvas.width - 10, originY - 10);
+  ctx.fillText("y", originX + 10, 10);
+
+  // draw x tick marks and labels
+  for (let x = -canvas.width / 2; x <= canvas.width / 2; x++) {
+    const xCoord = originX + x * scale;
+    ctx.moveTo(xCoord, originY - 5);
+    ctx.lineTo(xCoord, originY + 5);
+    ctx.stroke();
+    if (x !== 0) {
+      ctx.fillText(x.toString(), xCoord - 3, originY + 15);
+    } else {
+      ctx.fillText(x.toString(), xCoord + 5, originY + 15)
+    }
+  }
+
+  // draw y tick marks and labels
+  for (let y = -canvas.height / 2; y <= canvas.height / 2; y++) {
+    const yCoord = originY - y * scale;
+    ctx.moveTo(originX - 5, yCoord);
+    ctx.lineTo(originX + 5, yCoord);
+    ctx.stroke();
+    if (y !== 0) {
+      ctx.fillText(y.toString(), originX + 10, yCoord + 3);
+    }
+  }
+
+  // draw the graph
+  ctx.beginPath();
+  let x = -canvas.width * 2;
+  ctx.moveTo(-canvas.width * 2, -canvas.height * 2);
+  for (let i = -canvas.width / 2; i <= canvas.width / 2; i += step) {
+    let x = i;
+    if (isEvaluable(x)) {
+      ctx.lineTo(originX + x * scale, originY - eval(value) * scale);
+    }
+    console.log(`x = ${x} y = ${eval(value)}`)
+  }
+  ctx.stroke();
+}
