--- styles.css
@@ -1,78 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f5f5f5;
-    margin: 0;
-    padding: 20px;
-    margin-left: auto;
-}
-
-h1 {
-    color: #333;
-}
-
-input[type="text"] {
-    padding: 10px;
-    border-radius: 4px;
-    border: 1px solid #ccc;
-    width: 300px;
-    margin-right: 10px;
-}
-
-button {
-    margin-bottom: 10px;
-    margin-right: 10px;
-    padding: 10px 20px;
-    background-color: #4CAF50;
-    border: none;
-    color: white;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-button:hover {
-    background-color: #45a049;
-}
-
-ul {
-    list-style-type: none;
-    padding: 0;
-}
-
-li {
-    margin-bottom: 10px;
-}
-
-.listItem {
-    color: grey;
-    font-style: oblique;
-}
-
-canvas {
-    border: 1px solid #ccc;
-}
-
-.inputOfElementStyle {
-    padding: 10px;
-    border-radius: 4px;
-    border: 1px solid #ccc;
-    width: 300px;
-    margin-right: 10px;
-}
-
-.errorTextStyle {
-    font-size: 14px;
-    color: red;
-    font-style: italic;
-}
-
-.hidden {
-  display: none;
-  opacity: 0;
-  transition: opacity 0.3s ease;
-}
-
-.visible {
-  display: block;
-  opacity: 1;
-  transition: opacity 0.3s ease;
-}
